move config validation to validateConfigs.ts
