+++ src/sqlFormatter.ts
@@ -16,7 +16,7 @@ import SingleStoreDbFormatter from './languages/singlestoredb/singlestoredb.form
 import SnowflakeFormatter from './languages/snowflake/snowflake.formatter';
 
 import { FormatOptions } from './FormatOptions';
-import { ParamItems } from './formatter/Params';
+import { ConfigError, validateConfig, validateQuery } from './validateConfigs';
 
 export const formatters = {
   bigquery: BigQueryFormatter,
@@ -63,9 +63,7 @@ const defaultOptions: FormatOptions = {
  * @return {string} formatted query
  */
 export const format = (query: string, cfg: Partial<FormatOptions> = defaultOptions): string => {
-  if (typeof query !== 'string') {
-    throw new Error('Invalid query argument. Expected string, instead got ' + typeof query);
-  }
+  validateQuery(query);
 
   if (typeof cfg.language === 'string' && !supportedDialects.includes(cfg.language)) {
     throw new ConfigError(`Unsupported SQL dialect: ${cfg.language}`);
@@ -80,47 +78,6 @@ export const format = (query: string, cfg: Partial<FormatOptions> = defaultOptio
     typeof options.language === 'string' ? formatters[options.language] : options.language;
 
   return new FormatterCls(options).format(query);
-};
-
-export class ConfigError extends Error {}
-
-const validateConfig = (cfg: FormatOptions): FormatOptions => {
-  if ('multilineLists' in cfg) {
-    throw new ConfigError('multilineLists config is no more supported.');
-  }
-  if ('newlineBeforeOpenParen' in cfg) {
-    throw new ConfigError('newlineBeforeOpenParen config is no more supported.');
-  }
-  if ('newlineBeforeCloseParen' in cfg) {
-    throw new ConfigError('newlineBeforeCloseParen config is no more supported.');
-  }
-  if ('aliasAs' in cfg) {
-    throw new ConfigError('aliasAs config is no more supported.');
-  }
-
-  if (cfg.expressionWidth <= 0) {
-    throw new ConfigError(
-      `expressionWidth config must be positive number. Received ${cfg.expressionWidth} instead.`
-    );
-  }
-
-  if (cfg.commaPosition === 'before' && cfg.useTabs) {
-    throw new ConfigError(
-      'commaPosition: before does not work when tabs are used for indentation.'
-    );
-  }
-
-  if (cfg.params && !validateParams(cfg.params)) {
-    // eslint-disable-next-line no-console
-    console.warn('WARNING: All "params" option values should be strings.');
-  }
-
-  return cfg;
-};
-
-const validateParams = (params: ParamItems | string[]): boolean => {
-  const paramValues = params instanceof Array ? params : Object.values(params);
-  return paramValues.every(p => typeof p === 'string');
 };
 
 export type FormatFn = typeof format;
